requirements.txt
pymongo
mysql-connector-python
python-dotenv
mysqlclient